+++ Cargo.toml
@@ -12,7 +12,5 @@ env_logger = "0.9"
 fermium = { version = "20022.0.0", default-features = false }
 zstring = "0.1"
 log = "0.4"
-rayon = "1.5"
-parking_lot = "0.12.1"
-rand = "0.8.4"
 gridmath = { path = "gridmath" }
+sandworld = { path = "sandworld" }
