+++ sandworld/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "sandworld"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+gridmath = { path = "../gridmath" }
+rayon = "1.5"
+rand = "0.8.4"
