+++ sandworld/src/chunk.rs
@@ -1,17 +1,17 @@
+
+pub const CHUNK_SIZE: u8 = 64;
 use gridmath::*;
-use rand::{Rng, RngCore};
-use rayon::prelude::*;
-use std::{collections::HashMap, sync::atomic::AtomicU64};
+use rand::{Rng};
 
-const CHUNK_SIZE: u8 = 64;
-pub const WORLD_WIDTH: i32 = 1440;
-pub const WORLD_HEIGHT: i32 = 960;
+use crate::particle::*;
 
-pub struct World {
-    // TODO keep chunks at consistent addresses once allocated
-    // option - box all the chunks, store the boxes?
-    // option - static self managed array with map of coord to index
-    chunks: HashMap<u64, Box<Chunk>>,
+pub struct Chunk {
+    pub position: GridVec,
+    neighbors: Neighbors,
+    particles: [Particle; CHUNK_SIZE as usize * CHUNK_SIZE as usize],
+    pub(crate) dirty: Option<GridBounds>,
+    pub(crate) update_this_frame: Option<GridBounds>,
+    pub(crate) updated_last_frame: Option<GridBounds>,
 }
 
 struct Neighbors {
@@ -31,32 +31,6 @@ unsafe impl Sync for Neighbors {}
 struct NeighborIterator<'a> {
     neighbor_array: [&'a Option<*mut Chunk>; 8],
     cur_index: usize,
-}
-
-struct Chunk {
-    position: GridVec,
-    neighbors: Neighbors,
-    particles: [Particle; CHUNK_SIZE as usize * CHUNK_SIZE as usize],
-    dirty: Option<GridBounds>,
-    update_this_frame: Option<GridBounds>,
-    updated_last_frame: Option<GridBounds>,
-}
-
-#[derive(PartialEq, Debug, Copy, Clone)]
-pub enum ParticleType {
-    Air,
-    Sand,
-    Water,
-    Stone,
-    Source,
-    Boundary,
-    Dirty,
-}
-
-#[derive(Debug, Copy, Clone)]
-pub struct Particle {
-    pub particle_type: ParticleType,
-    updated_this_frame: bool,
 }
 
 impl Iterator for NeighborIterator<'_> {
@@ -114,33 +88,9 @@ impl Drop for Neighbors {
     }
 }
 
-impl Particle {
-    pub fn new(particle_type: ParticleType) -> Self {
-        Particle{particle_type, updated_this_frame: false}
-    }
-
-    fn get_possible_moves(particle_type: ParticleType) -> Vec::<GridVec> {
-        match particle_type {
-            ParticleType::Sand => vec![GridVec{x: 1, y: -1}, GridVec{x: -1, y: -1}, GridVec{x: 0, y: -1}],
-            ParticleType::Water => vec![GridVec{x: 1, y: -1}, GridVec{x: 0, y: -1}, GridVec{x: -1, y: -1}, GridVec{x: 1, y: 0}, GridVec{x: -1, y: 0} ],
-            _ => Vec::<GridVec>::new(),
-        }
-    }
-
-    fn can_replace_water(particle_type: ParticleType) -> bool {
-        match particle_type {
-            ParticleType::Sand => true,
-            _ => false,
-        }
-    }
-}
-
-impl Default for Particle {
-    fn default() -> Self { Particle{particle_type: ParticleType::Air, updated_this_frame: false} }
-}
 
 impl Chunk {
-    fn new(position: GridVec) -> Self {
+    pub fn new(position: GridVec) -> Self {
         let created = Chunk {
             position,
             neighbors: Neighbors::new(),
@@ -157,7 +107,7 @@ impl Chunk {
         return y as usize * CHUNK_SIZE as usize + x as usize;
     }
 
-    fn get_particle(&self, x: u8, y: u8) -> Particle {
+    pub fn get_particle(&self, x: u8, y: u8) -> Particle {
         #[cfg(debug_assertions)] {
             if x >= CHUNK_SIZE {
                 println!("X VALUE OF {} IS TOO LARGE", x);
@@ -172,7 +122,7 @@ impl Chunk {
         return self.particles[Chunk::get_index_in_chunk(x, y)];
     }
 
-    fn get_particle_mut(&mut self, x: u8, y: u8) -> &mut Particle {
+    pub fn get_particle_mut(&mut self, x: u8, y: u8) -> &mut Particle {
         #[cfg(debug_assertions)] {
             if x >= CHUNK_SIZE {
                 println!("X VALUE OF {} IS TOO LARGE", x);
@@ -185,11 +135,11 @@ impl Chunk {
         return &mut self.particles[Chunk::get_index_in_chunk(x, y)];
     }
 
-    fn contains(&self, x: i16, y: i16) -> bool {
+    pub fn contains(&self, x: i16, y: i16) -> bool {
         x >= 0 && y >= 0 && x < CHUNK_SIZE as i16 && y < CHUNK_SIZE as i16
     }
 
-    fn get_neighbor(&self, dir: GridVec) -> Option<*mut Chunk> {
+    pub fn get_neighbor(&self, dir: GridVec) -> Option<*mut Chunk> {
         if dir.y < 0 {
             if dir.x < 0 { 
                 return self.neighbors.bottom_left;
@@ -266,7 +216,7 @@ impl Chunk {
         return false;
     }
 
-    fn check_add_neighbor(&mut self, new_chunk: &mut Chunk) {
+    pub(crate) fn check_add_neighbor(&mut self, new_chunk: &mut Chunk) {
         if !self.position.is_adjacent(new_chunk.position) {
             return;
         }
@@ -313,18 +263,18 @@ impl Chunk {
         }
     }
 
-    fn chunkpos_to_local_chunkpos(&self, from_chunk: &Chunk, from_x: u8, from_y: u8) -> GridVec {
+    pub fn chunkpos_to_local_chunkpos(&self, from_chunk: &Chunk, from_x: u8, from_y: u8) -> GridVec {
         let chunk_diff = (from_chunk.position - self.position) * CHUNK_SIZE as i32;
         GridVec::new(from_x as i32, from_y as i32) + chunk_diff
     }
 
-    fn set_particle(&mut self, x: u8, y: u8, val: Particle) {
+    pub fn set_particle(&mut self, x: u8, y: u8, val: Particle) {
         self.particles[Chunk::get_index_in_chunk(x, y)] = val;
         self.mark_dirty(x as i32, y as i32);
         self.particles[Chunk::get_index_in_chunk(x, y)].updated_this_frame = true;
     }
 
-    fn mark_dirty(&mut self, x: i32, y: i32) {
+    pub fn mark_dirty(&mut self, x: i32, y: i32) {
         let chunk_bounds = GridBounds::new_from_corner(GridVec::new(0, 0), GridVec::new(CHUNK_SIZE as i32, CHUNK_SIZE as i32));
         let dirty_bounds = chunk_bounds.intersect(GridBounds::new(GridVec { x, y }, GridVec { x: 4, y: 4 }));
 
@@ -345,7 +295,7 @@ impl Chunk {
         }
     }
 
-    fn add_particle(&mut self, x: u8, y: u8, val: Particle) {
+    pub fn add_particle(&mut self, x: u8, y: u8, val: Particle) {
         if self.get_particle(x, y).particle_type == ParticleType::Air {
             self.particles[Chunk::get_index_in_chunk(x, y)] = val;
             self.mark_dirty(x as i32, y as i32);
@@ -356,7 +306,7 @@ impl Chunk {
         x == 0 || y == 0 || x == CHUNK_SIZE - 1  || y == CHUNK_SIZE - 1
     }
 
-    fn commit_updates(&mut self) {
+    pub(crate) fn commit_updates(&mut self) {
         self.update_this_frame = self.dirty;
         self.dirty = None;
 
@@ -370,7 +320,7 @@ impl Chunk {
         }
     }
 
-    fn update(&mut self) {      
+    pub(crate) fn update(&mut self) {      
         let mut rng = rand::thread_rng();
 
         if let Some(to_update) = GridBounds::option_union(self.update_this_frame, self.updated_last_frame) {
@@ -430,7 +380,7 @@ impl Chunk {
         self.updated_last_frame = self.update_this_frame;
     }
 
-    fn _check_remove_neighbor(&mut self, removed_position: GridVec) {
+    pub(crate) fn _check_remove_neighbor(&mut self, removed_position: GridVec) {
         if !self.position.is_adjacent(removed_position) {
             return;
         }
@@ -467,173 +417,5 @@ impl Chunk {
                 self.neighbors.top_right = None;
             }
         }
-    }
-}
-
-impl World {
-    pub fn new() -> Self {
-        let mut created: World = World {
-            chunks: HashMap::new(),
-        };
-
-        let world_width_chunks = (WORLD_WIDTH / CHUNK_SIZE as i32) + 1;
-        let world_height_chunks = (WORLD_HEIGHT / CHUNK_SIZE as i32) + 1;
-
-        for y in 0..world_height_chunks {
-            for x in 0..world_width_chunks {
-                let chunkpos = GridVec::new(x, y);
-                created.add_chunk(chunkpos);
-            }
-        }
-
-        return created;
-    }
-
-    pub fn contains(&self, pos: GridVec) -> bool {
-        let chunk_pos = World::get_chunkpos(pos);
-        return self.chunks.contains_key(&chunk_pos.combined());
-    }
-
-    fn add_chunk(&mut self, chunkpos: GridVec) {
-        let mut added = Box::new(Chunk::new(chunkpos));
-
-        for (_pos, chunk) in self.chunks.iter_mut() {
-            chunk.check_add_neighbor(&mut added);
-        }
-
-        self.chunks.insert(chunkpos.combined(), added);
-    }
-
-    fn get_chunkpos(pos: GridVec) -> GridVec {
-        GridVec::new(pos.x / CHUNK_SIZE as i32, pos.y / CHUNK_SIZE as i32)
-    }
-
-    fn get_chunklocal(pos: GridVec) -> GridVec {
-        let mut modded = GridVec::new(pos.x % CHUNK_SIZE as i32, pos.y % CHUNK_SIZE as i32);
-        if modded.x < 0 { 
-            modded.x += CHUNK_SIZE as i32; 
-        }
-        if modded.y < 0 { 
-            modded.y += CHUNK_SIZE as i32;
-        }
-        return modded;
-    }
-
-    pub fn _get_particle(&self, pos: GridVec) -> Particle {
-        if !self.contains(pos) {
-            return Particle::new(ParticleType::Boundary);
-        }
-
-        let chunk_pos = World::get_chunkpos(pos);
-        let chunklocal = World::get_chunklocal(pos);
-        return self.chunks.get(&chunk_pos.combined()).unwrap().get_particle(chunklocal.x as u8, chunklocal.y as u8);
-    }
-
-    pub fn replace_particle(&mut self, pos: GridVec, new_val: Particle) {
-        if !self.contains(pos) {
-            return;
-        }
-
-        let chunk_pos = World::get_chunkpos(pos);
-        let chunklocal = World::get_chunklocal(pos);
-        self.chunks.get_mut(&chunk_pos.combined()).unwrap().set_particle(chunklocal.x as u8, chunklocal.y as u8, new_val);
-    }
-
-    pub fn add_particle(&mut self, pos: GridVec, new_val: Particle) {
-        if !self.contains(pos) {
-            return;
-        }
-
-        let chunk_pos = World::get_chunkpos(pos);
-        let chunklocal = World::get_chunklocal(pos);
-        self.chunks.get_mut(&chunk_pos.combined()).unwrap().add_particle(chunklocal.x as u8, chunklocal.y as u8, new_val);
-    }
-
-    pub fn clear_circle(&mut self, pos: GridVec, radius: i32) {
-        self.place_circle(pos, radius, Particle::new(ParticleType::Air), true);
-    }
-
-    pub fn place_circle(&mut self, pos: GridVec, radius: i32, new_val: Particle, replace: bool) {
-        let left = pos.x - radius;
-        let right = pos.x + radius;
-        let bottom = pos.y - radius;
-        let top = pos.y + radius;
-
-        for y in bottom..top {
-            for x in left..right {
-                if replace { self.replace_particle(GridVec{x, y}, new_val.clone()); }
-                else { self.add_particle(GridVec{x, y}, new_val.clone()); }
-            }
-        }
-    }
-
-    pub fn render(&self, screen_bounds: &GridBounds, draw_debug: bool) -> Vec<Particle> {
-        let mut outbuffer = Vec::new();
-        let top_right = screen_bounds.top_right();
-        let bottom_left = screen_bounds.bottom_left();
-        let buffer_height = top_right.y - bottom_left.y;
-        let buffer_width = top_right.x - bottom_left.x;
-
-        outbuffer.resize(buffer_height as usize * buffer_width as usize, Particle::new(ParticleType::Boundary));
-
-        self.chunks.iter().for_each(|(chunk_pos_combined, chunk)| {
-            let chunk_pos = GridVec::decombined(*chunk_pos_combined);
-            let chunk_bottom_left = chunk_pos * CHUNK_SIZE as i32;
-            let chunk_bounds = GridBounds::new_from_corner(chunk_bottom_left, GridVec::new(CHUNK_SIZE as i32, CHUNK_SIZE as i32));
-
-            if let Some(overlap) = screen_bounds.intersect(chunk_bounds) {
-                for overlap_pos in overlap.iter() {
-                    let chunk_local = World::get_chunklocal(overlap_pos);
-                    let buffer_local = overlap_pos - bottom_left;
-                    let buffer_index = (buffer_local.x + (buffer_width * buffer_local.y)) as usize;
-
-                    let mut write_val = chunk.get_particle(chunk_local.x as u8, chunk_local.y as u8);
-
-                    if draw_debug {
-                        if chunk_local.x == 0 || chunk_local.y == 0 || chunk_local.x as u8 == CHUNK_SIZE - 1 || chunk_local.y as u8 == CHUNK_SIZE - 1 {
-                            write_val = Particle::new(ParticleType::Boundary);
-                        }
-                        if let Some(updated_bounds) = chunk.updated_last_frame {
-                            if updated_bounds.is_boundary(chunk_local){
-                                write_val = Particle::new(ParticleType::Dirty);
-                            }
-                        }
-                    }
-
-                    outbuffer[buffer_index] = write_val;
-                }
-            }
-        });
-
-        return outbuffer;
-    }
-
-    pub fn update(&mut self) -> u64 {
-        let updated_count = AtomicU64::new(0);
-
-        self.chunks.par_iter_mut().for_each(|(_pos, chunk)| {
-            chunk.commit_updates();
-        });
-
-        let shift = (rand::thread_rng().next_u32() % 4) as i32;
-
-        for i in 0..4{
-            
-            let x_mod = (i + shift) % 2;
-            let y_mod = ((i + shift) / 2) % 2; 
-
-            self.chunks.par_iter_mut().for_each(|(pos, chunk)| {
-                let chunk_pos = GridVec::decombined(*pos);
-
-                if chunk_pos.x % 2 == x_mod && chunk_pos.y % 2 == y_mod {
-                    if chunk.update_this_frame.is_some() || chunk.updated_last_frame.is_some() { 
-                        updated_count.fetch_add(1, std::sync::atomic::Ordering::Relaxed);
-                        chunk.update(); 
-                    }
-                }
-            });
-        }
-        
-        updated_count.load(std::sync::atomic::Ordering::Relaxed)
     }
 }
