+++ sandworld/src/lib.rs
@@ -1,0 +1,8 @@
+mod particle;
+mod chunk;
+mod sandworld;
+
+pub use sandworld::*;
+pub use particle::*;
+pub use chunk::*;
+
